requirements.txt
numpy
opencv-python
mediapipe
pyrealsense2
RPi.GPIO